requirements.txt
numpy>=1.24.0
scipy>=1.10.0
librosa>=0.10.0
opencv-python>=4.8.0
matplotlib>=3.7.0
mutagen>=1.47.0
pillow>=10.0.0
piexif>=1.1.3
pyexiftool>=0.5.5
pydub>=0.25.1
pandas>=2.0.0
scikit-learn>=1.3.0
pytest>=7.4.0
jupyter>=1.0.0
